requirements.txt
Flask
Pillow
dash
pandas
dash-table
plotly
requests